Use sqlalchemy sessions correctly so locking works
Fixes bug 990118

When using with_lockmode, it's necessary to use the same session for both
the select and the subsequent update/delete. Since a new session was
being used for each part, the lock was becoming an effective noop and
two requests could race with each other. Make sure we use one session
to prevent this from happening.

Change-Id: I6765c97266d13b795f1153b440cfc556df065787